backend/requirements.txt
# Core FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# RAG & AI
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.10
langchain-google-genai==2.0.0
chromadb==0.4.22
sentence-transformers==2.2.2
anthropic==0.18.1
cohere==5.0.0
google-genai==0.3.0

# CV Processing
mistralai==1.0.0
groq==0.11.0

# Gmail Integration
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0

# Utilities
pydantic[email]==2.5.3
pydantic-settings==2.1.0
